+++ ChangeLog.md
@@ -1,3 +1,15 @@
+### 2.0.18
+
+* ToJavascript instance for String, Strict and Lazy Text [#227](https://github.com/yesodweb/shakespeare/pull/227)
+
+### 2.0.17
+
+* Fix parse pattern-match with operator constructor [#222](https://github.com/yesodweb/shakespeare/issues/222)
+
+### 2.0.16
+
+* Updated `encodeToTextBuilder` also escapes single quotes [#221](https://github.com/yesodweb/shakespeare/pull/221)
+
 ### 2.0.15
 
 * Semigroup instances